requirements.txt
asgiref==3.9.1
boto3==1.40.4
botocore==1.40.4
Django==5.2.4
djangorestframework==3.16.0
flake8==7.3.0
iniconfig==2.1.0
jmespath==1.0.1
mccabe==0.7.0
packaging==25.0
pluggy==1.6.0
psycopg==3.2.9
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
